remove remove pn_improved
