+++ skglm/solvers/pn_solver.py
@@ -1,17 +1,20 @@
 import numpy as np
-from scipy.sparse import issparse
 from numba import njit
+from scipy.sparse import issparse
 
-from skglm.solvers.common import construct_grad, construct_grad_sparse
+
+EPS_TOL = 0.3
+MAX_CD_ITER = 20
+MAX_BACKTRACK_ITER = 20
 
 
-def pn_solver(X, y, datafit, penalty, w_init=None, p0=10,
-              max_iter=20, max_epochs=1000, tol=1e-4, verbose=0):
+def prox_newton(X, y, datafit, penalty, w_init=None, p0=10,
+                max_iter=20, max_epochs=1000, tol=1e-4, verbose=0):
     """Run a Prox Newton solver.
 
     Parameters
     ----------
-    X : array, shape (n_samples, n_features)
+    X : array or sparse CSC matrix shape (n_samples, n_features)
         Design matrix.
 
     y : array, shape (n_samples,)
@@ -28,10 +31,10 @@ def pn_solver(X, y, datafit, penalty, w_init=None, p0=10,
         If set to None, a zero vector is used instead.
 
     p0 : int, default 10
-        Minimum number of groups to be included in the working set.
+        Minimum number of features to be included in the working set.
 
     max_iter : int, default 20
-        Maximum number of iterations.
+        Maximum number of outer iterations.
 
     max_epochs : int, default 1000
         Maximum number of epochs.
@@ -51,7 +54,7 @@ def pn_solver(X, y, datafit, penalty, w_init=None, p0=10,
         The objective values at every outer iteration.
 
     stop_crit: float
-        The value of the stop criterion.
+        The value of the stopping criterion when the solver stops.
     """
     n_samples, n_features = X.shape
     w = np.zeros(n_features) if w_init is None else w_init
@@ -93,36 +96,29 @@ def pn_solver(X, y, datafit, penalty, w_init=None, p0=10,
                       min(n_features, 2 * gsupp_size))
         # similar to np.argsort()[-ws_size:] but without sorting
         ws = np.argpartition(opt, -ws_size)[-ws_size:]
+
         grad_ws = grad[ws]
+        tol_in = EPS_TOL * stop_crit
 
         for epoch in range(max_epochs):
-            tol_in = 0.3 * stop_crit
 
             # find descent direction
             if is_sparse:
-                (
-                    delta_w_ws,
-                    X_delta_w_ws
-                ) = _compute_descent_direction_s(*X_bundles, y, w, Xw, grad_ws,
-                                                 datafit, penalty, ws,
-                                                 max_cd_iter=20, tol=0.3*tol_in)
+                delta_w_ws, X_delta_w_ws = _compute_descent_direction_s(
+                    *X_bundles, y, w, Xw, grad_ws, datafit,
+                    penalty, ws, tol=EPS_TOL*tol_in)
             else:
-                (
-                    delta_w_ws,
-                    X_delta_w_ws
-                ) = _compute_descent_direction(X, y, w, Xw, grad_ws, datafit, penalty,
-                                               ws, max_cd_iter=20, tol=0.3*tol_in)
+                delta_w_ws, X_delta_w_ws = _compute_descent_direction(
+                    X, y, w, Xw, grad_ws, datafit, penalty, ws, tol=EPS_TOL*tol_in)
 
             # backtracking line search with inplace update of w, Xw
             if is_sparse:
-                grad_ws[:] = _backtrack_line_search_s(*X_bundles, y, w, Xw, datafit,
-                                                      penalty, delta_w_ws,
-                                                      X_delta_w_ws, ws,
-                                                      max_backtrack_iter=20)
+                grad_ws[:] = _backtrack_line_search_s(*X_bundles, y, w, Xw,
+                                                      datafit, penalty, delta_w_ws,
+                                                      X_delta_w_ws, ws)
             else:
                 grad_ws[:] = _backtrack_line_search(X, y, w, Xw, datafit, penalty,
-                                                    delta_w_ws, X_delta_w_ws, ws,
-                                                    max_backtrack_iter=20)
+                                                    delta_w_ws, X_delta_w_ws, ws)
 
             # check convergence
             opt_in = penalty.subdiff_distance(w, grad_ws, ws)
@@ -146,12 +142,12 @@ def pn_solver(X, y, datafit, penalty, w_init=None, p0=10,
 
 
 @njit
-def _compute_descent_direction(X, y, w_epoch, Xw_epoch, grad_ws, datafit, penalty,
-                               ws, max_cd_iter, tol):
+def _compute_descent_direction(X, y, w_epoch, Xw_epoch, grad_ws, datafit,
+                               penalty, ws, tol):
     # Given:
-    #   - b = \nabla   F(X w_epoch)
-    #   - D = \nabla^2 F(X w_epoch)   <------>  raw_hess
-    # Minimize for delta_w = w - w_epoch:
+    #   1) b = \nabla F(X w_epoch)
+    #   2) D = \nabla^2 F(X w_epoch)  <------>  raw_hess
+    # Minimize quadratic approximation for delta_w = w - w_epoch:
     #  b.T @ X @ delta_w + \
     #  1/2 * delta_w.T @ (X.T @ D @ X) @ delta_w + penalty(w)
     raw_hess = datafit.raw_hessian(y, Xw_epoch)
@@ -164,9 +160,8 @@ def _compute_descent_direction(X, y, w_epoch, Xw_epoch, grad_ws, datafit, penalt
     X_delta_w_ws = np.zeros(X.shape[0])
     w_ws = w_epoch[ws]
 
-    for cd_iter in range(max_cd_iter):
+    for cd_iter in range(MAX_CD_ITER):
         for idx, j in enumerate(ws):
-
             # skip when X[:, j] == 0
             if lipschitz[idx] == 0:
                 continue
@@ -176,9 +171,7 @@ def _compute_descent_direction(X, y, w_epoch, Xw_epoch, grad_ws, datafit, penalt
             stepsize = 1 / lipschitz[idx]
 
             w_ws[idx] = penalty.prox_1d(
-                old_w_idx - stepsize * cached_grads[idx],
-                stepsize, j
-            )
+                old_w_idx - stepsize * cached_grads[idx], stepsize, j)
 
             if w_ws[idx] != old_w_idx:
                 X_delta_w_ws += (w_ws[idx] - old_w_idx) * X[:, j]
@@ -192,45 +185,42 @@ def _compute_descent_direction(X, y, w_epoch, Xw_epoch, grad_ws, datafit, penalt
     return w_ws - w_epoch[ws], X_delta_w_ws
 
 
-# sparse version of func above
+# sparse version of _compute_descent_direction
 @njit
-def _compute_descent_direction_s(X_data, X_indptr, X_indices, y,
-                                 w_epoch, Xw_epoch, grad_ws, datafit, penalty,
-                                 ws, max_cd_iter, tol):
+def _compute_descent_direction_s(X_data, X_indptr, X_indices, y, w_epoch,
+                                 Xw_epoch, grad_ws, datafit, penalty, ws, tol):
     raw_hess = datafit.raw_hessian(y, Xw_epoch)
 
     lipschitz = np.zeros(len(ws))
     for idx, j in enumerate(ws):
-        for i in range(X_indptr[j], X_indptr[j+1]):
-            lipschitz[idx] += raw_hess[X_indices[i]] * X_data[i] ** 2
+        # equivalent to: lipschitz[idx] += raw_hess * X[:, j] ** 2
+        lipschitz[idx] = sparse_squared_weighted_norm(X_data, X_indptr, X_indices,
+                                                      j, raw_hess)
 
     cached_grads = np.zeros(len(ws))
     X_delta_w_ws = np.zeros(len(y))
     w_ws = w_epoch[ws]
 
-    for cd_iter in range(max_cd_iter):
+    for cd_iter in range(MAX_CD_ITER):
         for idx, j in enumerate(ws):
-
             # skip when X[:, j] == 0
             if lipschitz[idx] == 0:
                 continue
 
             cached_grads[idx] = grad_ws[idx]
-            for i in range(X_indptr[j], X_indptr[j+1]):
-                row_i = X_indices[i]
-                cached_grads[idx] += X_data[i] * raw_hess[row_i] * X_delta_w_ws[row_i]
+            # equivalent to cached_grads[idx] += X[:, j] @ (raw_hess * X_delta_w_ws)
+            cached_grads[idx] += sparse_weighted_dot(X_data, X_indptr, X_indices, j,
+                                                     X_delta_w_ws, raw_hess)
 
             old_w_idx = w_ws[idx]
             stepsize = 1 / lipschitz[idx]
 
             w_ws[idx] = penalty.prox_1d(
-                old_w_idx - stepsize * cached_grads[idx],
-                stepsize, j
-            )
+                old_w_idx - stepsize * cached_grads[idx], stepsize, j)
 
             if w_ws[idx] != old_w_idx:
-                for i in range(X_indptr[j], X_indptr[j+1]):
-                    X_delta_w_ws[X_indices[i]] += (w_ws[idx] - old_w_idx) * X_data[i]
+                update_X_delta_w(X_data, X_indptr, X_indices, X_delta_w_ws,
+                                 w_ws[idx] - old_w_idx, j)
 
         if cd_iter % 5 == 0:
             opt = penalty.subdiff_distance(w_ws, cached_grads, ws)
@@ -242,14 +232,14 @@ def _compute_descent_direction_s(X_data, X_indptr, X_indices, y,
 
 
 @njit
-def _backtrack_line_search(X, y, w, Xw, datafit, penalty, delta_w_ws, X_delta_w_ws,
-                           ws, max_backtrack_iter):
+def _backtrack_line_search(X, y, w, Xw, datafit, penalty, delta_w_ws,
+                           X_delta_w_ws, ws):
     # inplace update of w and Xw
     # return grad_ws of the last w and Xw
     step, prev_step = 1., 0.
     prev_penalty_val = penalty.value(w[ws])
 
-    for backtrack_iter in range(max_backtrack_iter):
+    for backtrack_iter in range(MAX_BACKTRACK_ITER):
         stop_crit = -prev_penalty_val
         w[ws] += (step - prev_step) * delta_w_ws
         Xw += (step - prev_step) * X_delta_w_ws
@@ -267,17 +257,16 @@ def _backtrack_line_search(X, y, w, Xw, datafit, penalty, delta_w_ws, X_delta_w_
     return grad_ws
 
 
-# sparse version of func above
+# sparse version of _backtrack_line_search
 @njit
-def _backtrack_line_search_s(X_data, X_indptr, X_indices,
-                             y, w, Xw, datafit, penalty, delta_w_ws,
-                             X_delta_w_ws, ws, max_backtrack_iter):
+def _backtrack_line_search_s(X_data, X_indptr, X_indices, y, w, Xw, datafit,
+                             penalty, delta_w_ws, X_delta_w_ws, ws):
     # inplace update of w and Xw
     step, prev_step = 1., 0.
-    prev_penalty_val = penalty.value(w[ws])
+    epoch_penalty_val = penalty.value(w[ws])
 
-    for backtrack_iter in range(max_backtrack_iter):
-        stop_crit = -prev_penalty_val
+    for backtrack_iter in range(MAX_BACKTRACK_ITER):
+        stop_crit = -epoch_penalty_val
         w[ws] += (step - prev_step) * delta_w_ws
         Xw += (step - prev_step) * X_delta_w_ws
 
@@ -293,3 +282,57 @@ def _backtrack_line_search_s(X_data, X_indptr, X_indices,
             step /= 2
 
     return grad_ws
+
+
+@njit
+def construct_grad(X, y, w, Xw, datafit, ws):
+    """Compute grad of datafit restricted to ``ws``."""
+    raw_grad = datafit.raw_grad(y, Xw)
+    grad = np.zeros(len(ws))
+    for idx, j in enumerate(ws):
+        grad[idx] = X[:, j] @ raw_grad
+    return grad
+
+
+@njit
+def construct_grad_sparse(X_data, X_indptr, X_indices, y, w, Xw, datafit, ws):
+    """Compute grad of datafit restricted to ``ws`` in case ``X`` sparse."""
+    raw_grad = datafit.raw_grad(y, Xw)
+    grad = np.zeros(len(ws))
+    for idx, j in enumerate(ws):
+        grad[idx] = sparse_xj_dot(X_data, X_indptr, X_indices, j, raw_grad)
+    return grad
+
+
+@njit(fastmath=True)
+def sparse_xj_dot(X_data, X_indptr, X_indices, j, other):
+    """Compute ``X[:, j] @ other`` in case ``X`` sparse."""
+    res = 0.
+    for i in range(X_indptr[j], X_indptr[j+1]):
+        res += X_data[i] * other[X_indices[i]]
+    return res
+
+
+@njit(fastmath=True)
+def sparse_weighted_dot(X_data, X_indptr, X_indices, j, other, weights):
+    """Compute ``X[:, j] @ (weights * other)`` in case ``X`` sparse."""
+    res = 0.
+    for i in range(X_indptr[j], X_indptr[j+1]):
+        res += X_data[i] * other[X_indices[i]] * weights[X_indices[i]]
+    return res
+
+
+@njit(fastmath=True)
+def sparse_squared_weighted_norm(X_data, X_indptr, X_indices, j, weights):
+    """Compute ``weights @ X[:, j]**2`` in case ``X`` sparse."""
+    res = 0.
+    for i in range(X_indptr[j], X_indptr[j+1]):
+        res += weights[X_indices[i]] * X_data[i]**2
+    return res
+
+
+@njit(fastmath=True)
+def update_X_delta_w(X_data, X_indptr, X_indices, X_delta_w, diff, j):
+    """Compute ``X_delta_w += diff * X[:, j]`` case of ``X`` sparse."""
+    for i in range(X_indptr[j], X_indptr[j+1]):
+        X_delta_w[X_indices[i]] += diff * X_data[i]
